requirements.txt
backports.lzma==0.0.13
beautifulsoup4==4.7.1
CacheControl==0.12.5
certifi==2019.3.9
chardet==3.0.4
idna==2.8
lockfile==0.12.2
lxml==4.3.3
msgpack==0.6.1
nose==1.3.7
requests==2.21.0
soupsieve==1.9.1
urllib3==1.24.3
